Process programming
/* chainwaitmsg.c--
	The following program creates a process chain. Only one forked process
is a child of the original process. How are the output messages ordered?

   AUTHOR:Ruikang Dai
   DATE:Sat Mar 14 17:39:36 EDT 2015
*/
/* fanwaitmsg.c--
	The following program creates a process fan. All the forked processes
are children of the original process. How are the output messages
ordered??
	Because none of the forked children are parents, their wait function
returns –1 and sets errno to ECHILD. They are not blocked by the second
for loop. Their identification messages may appear in any order. The
message from the original process comes out at the very end after it
has waited for all of its children.
   AUTHOR:Ruikang Dai
   DATE:Sat Mar 14 17:29:01 EDT 2015
*/
/* parentwaitpid.c--
	Describe the possible forms of the output from the following program:
	Because none of the forked children are parents, their wait function
returns –1 and sets errno to ECHILD. They are not blocked by the second
for loop. Their identification messages may appear in any order. The
message from the original process comes out at the very end after it
has waited for all of its children.
   AUTHOR:Ruikang Dai
   DATE:Sat Mar 14 17:21:58 EDT 2015
*/
/* simplechain.c--
	A program that creates a chain of "n" processes, where n is a
command-line argument.
   AUTHOR:Ruikang Dai
   DATE:Sat Mar 14 15:20:24 EDT 2015
*/
/* simplefork.c--
	In the following program, both parent and child execute the x=1
assignment statement after running from fork.
   AUTHOR:Ruikang Dai
   DATE:Sat Mar 14 15:07:15 EDT 2015
*/